# backend/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  app.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
  app.ClusterInput:
    properties:
      name:
        type: string
      nodes:
        items:
          $ref: '#/definitions/app.NodeInput'
        type: array
      token:
        type: string
    required:
    - name
    - nodes
    type: object
...
  app.DeploymentListResponse:
    properties:
      count:
        type: integer
      deployments:
        items:
          $ref: '#/definitions/app.DeploymentResponse'
        type: array
    type: object
  app.DeploymentResponse:
    properties:
      cluster: {}
      created_at:
        type: string
      id:
        type: integer
      project_name:
        type: string
      updated_at:
        type: string
    type: object
  app.DeploymentWorkflowResponse:
    properties:
      status:
        type: string
      task_id:
        type: string
    type: object
...
  app.KubeconfigResponse:
    properties:
      kubeconfig:
        type: string
    type: object
...
  app.NodeInput:
    properties:
      cpu:
        type: integer
      disk_size:
        description: Storage in MB
        type: integer
      entrypoint:
        type: string
      env_vars:
        additionalProperties:
          type: string
        description: SSH_KEY, etc.
        type: object
      flist:
        type: string
      gpu_ids:
        description: List of GPU IDs
        items:
          type: string
        type: array
      memory:
        description: Memory in MB
        type: integer
      name:
        type: string
      node_id:
        type: integer
      root_size:
        description: Storage in MB
        type: integer
      type:
        enum:
        - worker
        - master
        - leader
        type: string
    required:
    - cpu
    - memory
    - name
    - node_id
    - root_size
    - type
    type: object
...
info:
  contact: {}
  description: API documentation for Mycelium Cloud.
  title: Mycelium Cloud API
  version: "1.0"
paths:
  /deployments:
    delete:
      description: Deletes all deployments and their resources for the authenticated
        user
      produces:
      - application/json
      responses:
        "202":
          description: Delete all deployments workflow started successfully
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/app.DeploymentWorkflowResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete all deployments
      tags:
      - deployments
    get:
      description: Retrieves a list of all deployments (clusters) for the authenticated
        user
      produces:
      - application/json
      responses:
        "200":
          description: Deployments retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/app.DeploymentListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - BearerAuth: []
      summary: List deployments
      tags:
      - deployments
    post:
      consumes:
      - application/json
      description: Creates and deploys a new Kubernetes cluster
      parameters:
      - description: Cluster configuration
        in: body
        name: cluster
        required: true
        schema:
          $ref: '#/definitions/app.ClusterInput'
      produces:
      - application/json
      responses:
        "202":
          description: Deployment workflow started successfully
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/app.DeploymentWorkflowResponse'
              type: object
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/app.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - BearerAuth: []
      summary: Deploy cluster
      tags:
      - deployments
  /deployments/{name}:
    delete:
      description: Deletes a specific deployment and all its resources
      parameters:
      - description: Deployment name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Deployment deletion workflow started successfully
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/app.DeploymentWorkflowResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/app.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.APIResponse'
        "404":
          description: Deployment not found
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete deployment
      tags:
      - deployments
    get:
      description: Retrieves details of a specific deployment by name
      parameters:
      - description: Deployment name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deployment details retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/app.DeploymentResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/app.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.APIResponse'
        "404":
          description: Deployment not found
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - BearerAuth: []
      summary: Get deployment
      tags:
      - deployments
  /deployments/{name}/kubeconfig:
    get:
      description: Retrieves the kubeconfig file for a specific deployment
      parameters:
      - description: Deployment name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Kubeconfig retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/app.KubeconfigResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/app.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.APIResponse'
        "404":
          description: Deployment not found
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - BearerAuth: []
      summary: Get kubeconfig
      tags:
      - deployments
  /deployments/{name}/nodes:
    post:
      consumes:
      - application/json
      description: Adds a new node to an existing deployment
      parameters:
      - description: Cluster configuration with new node
        in: body
        name: cluster
        required: true
        schema:
          $ref: '#/definitions/app.ClusterInput'
      produces:
      - application/json
      responses:
        "202":
          description: Node addition workflow started successfully
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/app.DeploymentWorkflowResponse'
              type: object
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/app.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.APIResponse'
        "404":
          description: Deployment not found
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - BearerAuth: []
      summary: Add node to deployment
      tags:
      - deployments
  /deployments/{name}/nodes/{node_name}:
    delete:
      description: Removes a specific node from an existing deployment
      parameters:
      - description: Deployment name
        in: path
        name: name
        required: true
        type: string
      - description: Node name to remove
        in: path
        name: node_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Node removal workflow started successfully
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/app.DeploymentWorkflowResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/app.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.APIResponse'
        "404":
          description: Deployment not found
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - BearerAuth: []
      summary: Remove node from deployment
      tags:
      - deployments
...
swagger: "2.0"